requirements.txt
python-jenkins>=1.1.0
